Client class updates
Client class is updated as requested.
The clientId is generated by database, when inserting an entry to the "client" table (assumed it's already & primary key autoincrement).
The last id just entered from table would be assigned to clientId for the client object. so that to avoid same id appears when server gets restarted.